+++ app/controllers/system/groups_controller.rb
@@ -1,6 +1,8 @@
 class System::GroupsController < ApplicationController
+  include PublicAccessibleController
+
   def index
-    render json: Grouping::Group.all, root: 'groups'
+    show_all_groups
   end
 
   def show
@@ -14,12 +16,14 @@ class System::GroupsController < ApplicationController
   end
 
   def update
+    return render_failure unless group.can_edit?
     group.name = params[:name]
     group.save!
     render_success
   end
 
   def destroy
+    return render_failure unless group.can_edit?
     group.destroy
     render_success
   end
